backend/requirements.txt
fastapi
uvicorn
pydantic
passlib[bcrypt]
pyjwt
python-multipart
sentence-transformers
chromadb
numpy
tqdm
pdfplumber
python-docx
requests
python-dotenv
sqlmodel
sqlalchemy
aiosqlite


